kafka_api/requirements.txt
aiokafka==0.8.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
backoff==2.2.1
certifi==2023.5.7
click==8.1.4
clickhouse-driver==0.2.6
fastapi==0.100.0
flake8==6.0.0
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
kafka-python==2.0.2
mccabe==0.7.0
orjson==3.9.2
packaging==23.1
pycodestyle==2.10.0
pydantic==2.0.2
pydantic-settings==2.0.1
pydantic_core==2.1.2
pyflakes==3.0.1
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzlocal==5.0.1
uvicorn==0.22.0